Tabular Playground Series - Apr 2021/requirements.txt
ipyparallel
jupyter
matplotlib
neptune-client
neptune-notebooks
numpy
pandas
pandas_profiling
ppscore
psutil
scikit-learn
seaborn
xgboost

